Added the -Wconversion flag to CMake builds (#518)
* initial fixes for conversion errors

* Fixed remaining wConversion warnings

* Fixed build errors

* Added compiler definitions to prevent unit test builds from blocking Wconversion inclusion

* Fixing windows conversion error that linux didn't catch

+++ src/agent/device_info_interface/src/device_info_interface.c
@@ -223,7 +223,7 @@ void DeviceInfoInterface_ReportChangedPropertiesAsync()
                 Log_Error("Cannot convert property value to number. Value: %s", propertyValue);
                 goto done;
             }
-            json_object_set_number(root_object, propertyName, val);
+            json_object_set_number(root_object, propertyName, (double)val);
         }
     }
